+++ src/main/java/com/aivle/carekids/domain/user/general/jwt/constants/JwtUtils.java
@@ -134,7 +134,7 @@ public class JwtUtils {
     }
 
     private boolean isLogout(String accessToken) {
-        Long isLogout = jwtRepository.getValues(accessToken);
+        Integer isLogout = jwtRepository.getValues(accessToken);
 
         return isLogout == -1;
     }
